requirements.txt
pandas
psycopg2
sqlalchemy
google-genai
streamlit
chromadb
sentence-transformers
plotly